# docs/api/swagger.yaml
definitions:
  metrics.MetricExchange:
    properties:
      delta:
        type: integer
      id:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
info:
  contact:
    email: evshuvalov@yandex.ru
    name: Evgeniy Shuvalov
  description: Service for storing metrics data.
  title: Metrics collector API
  version: "1.0"
paths:
  /:
    get:
      operationId: homepage
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Yet another homepage
      tags:
      - Metrics
  /ping:
    get:
      operationId: health_info
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
        "501":
          description: Not Implemented
          schema:
            type: string
      summary: Verify server up and running
      tags:
      - Healthcheck
  /update:
    post:
      consumes:
      - application/json
      operationId: metrics_json_update
      parameters:
      - description: Request parameters.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metrics.MetricExchange'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/metrics.MetricExchange'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Push metric data as JSON
      tags:
      - Metrics
  /update/{type}/{name}/{value}:
    post:
      operationId: metrics_update
      parameters:
      - description: Metrics type (e.g. `counter`, `gauge`).
        in: path
        name: type
        required: true
        type: string
      - description: Metrics name.
        in: path
        name: name
        required: true
        type: string
      - description: Metrics value, must be convertable to `int64` or `float64`.
        in: path
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Push metric data.
      tags:
      - Metrics
  /updates:
    post:
      consumes:
      - application/json
      operationId: metrics_json_update_list
      parameters:
      - description: List of metrics to update.
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/metrics.MetricExchange'
          type: array
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/metrics.MetricExchange'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Push list of metrics data as JSON
      tags:
      - Metrics
  /value:
    post:
      consumes:
      - application/json
      operationId: metrics_json_info
      parameters:
      - description: 'Request parameters: `id` and `type` are required.'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metrics.MetricExchange'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/metrics.MetricExchange'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get metrics value as JSON
      tags:
      - Metrics
  /value/{type}/{name}:
    get:
      operationId: metrics_info
      parameters:
      - description: Metrics type (e.g. `counter`, `gauge`).
        in: path
        name: type
        required: true
        type: string
      - description: Metrics name.
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get metric's value as string
      tags:
      - Metrics
swagger: "2.0"
tags:
- description: '"Metrics API"'
  name: Metrics
- description: '"API to inspect service health state"'
  name: Healthcheck